requirements.txt
Flask==0.11.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.14
Tempita==0.5.2
WTForms==2.1
WTForms-Components==0.10.0
Werkzeug==0.11.10
click==6.6
decorator==4.0.10
infinity==1.4
intervals==0.8.0
itsdangerous==0.24
pbr==1.10.0
six==1.10.0
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
validators==0.10.3
